Moved SampleModule from src to app
