+++ app/SampleModule/SampleModuleCtrl.ts
